SuperElonAI/requirements.txt
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==3.7.1
astroid==3.3.10
attrs==25.3.0
autopep8==2.3.2
bcrypt==4.3.0
beautifulsoup4==4.13.4
black==25.1.0
blake3==1.0.5
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
deep-translator==1.11.4
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.4.0
distlib==0.3.9
distro==1.9.0
filelock==3.18.0
flake8==7.2.0
Flask==3.1.1
fonttools==4.58.1
fpdf2==2.8.3
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.12.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.15.0
httplib2==0.22.0
httpx==0.23.0
hyperframe==5.2.0
idna==2.10
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kaleido==0.2.1
kiwisolver==1.4.8
lazy-object-proxy==1.11.0
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mccabe==0.7.0
mypy_extensions==1.1.0
mysql-connector-python==9.3.0
narwhals==1.41.0
networkx==3.5
numpy==2.2.6
openai>=1.0.0
google-generativeai>=0.8.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pefile==2023.2.7
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.2
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
pyflakes==3.3.2
PyGithub==2.6.1
pygraphviz==1.14
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
PyJWT==2.10.1
pylint==3.3.7
PyMuPDF==1.26.0
PyMySQL==1.1.1
PyNaCl==1.5.0
pyparsing==3.2.3
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
pytz==2025.2
pywifi==1.1.12
pywin32-ctypes==0.2.3
referencing==0.36.2
requests==2.32.3
rfc3986==1.5.0
rpds-py==0.25.1
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
streamlit==1.45.1
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.3
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
uv==0.7.13
virtualenv==20.31.2
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
